Get network requests from DevTools directly and a few tweaks.
1. Add a README.
2. Improve URL classification for ad requests and impressions.
3. Tweak config.

DevTools is used instead of  Lighthouse's load data since it seems to be
missing some records. I noticed this on a few specific sites where
GPT and ad requests are missing. This warrants further investigation and may be
an issue within lighthouse. I made this change
to make sure the data is accurate in the meantime.

Change-Id: I5990199a345cd298fe919e669a0dcc4d539bfd01